# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResolvePushConflictRequest": {
        "description": "Resolve a concurrent-edit push (#379): the pushed content always lands as\na new fork; ``resolution`` only chooses who is primary (main/fork).",
        "properties": {
          "baseId": {
            "minLength": 1,
            "title": "Baseid",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional new display name for the fork",
            "title": "Name"
          },
          "resolution": {
            "enum": [
              "main",
              "fork"
            ],
            "title": "Resolution",
            "type": "string"
          },
          "sourcePath": {
            "minLength": 1,
            "title": "Sourcepath",
            "type": "string"
          }
        },
        "required": [
          "sourcePath",
          "baseId",
          "resolution"
        ],
        "title": "ResolvePushConflictRequest",
        "type": "object"
      },
...
      "ScanOptionsRequest": {
        "properties": {
          "awsProfile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awsprofile"
          },
          "awsRegion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awsregion"
          },
          "awsSessionToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awssessiontoken"
          },
          "llmApiKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llmapikey"
          },
          "llmApiVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llmapiversion"
          },
          "llmBaseUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llmbaseurl"
          },
          "llmConsensusRuns": {
            "default": 1,
            "title": "Llmconsensusruns",
            "type": "integer"
          },
          "llmMaxTokens": {
            "default": 8192,
            "title": "Llmmaxtokens",
            "type": "integer"
          },
          "llmModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llmmodel"
          },
          "llmProvider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llmprovider"
          },
          "useLlm": {
            "default": true,
            "title": "Usellm",
            "type": "boolean"
          }
        },
        "title": "ScanOptionsRequest",
        "type": "object"
      },
      "ScanResultResponse": {
        "properties": {
          "analyzersUsed": {
            "items": {
              "type": "string"
            },
            "title": "Analyzersused",
            "type": "array"
          },
          "durationSeconds": {
            "title": "Durationseconds",
            "type": "number"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ScanFindingResponse"
            },
            "title": "Findings",
            "type": "array"
          },
          "findingsCount": {
            "title": "Findingscount",
            "type": "integer"
          },
          "isSafe": {
            "title": "Issafe",
            "type": "boolean"
          },
          "maxSeverity": {
            "title": "Maxseverity",
            "type": "string"
          },
          "skillName": {
            "title": "Skillname",
            "type": "string"
          }
        },
        "required": [
          "skillName",
          "isSafe",
          "maxSeverity",
          "findingsCount",
          "findings",
          "analyzersUsed",
          "durationSeconds"
        ],
        "title": "ScanResultResponse",
        "type": "object"
      },
...
      "SkillDetailResponse": {
        "properties": {
          "actions": {
            "$ref": "#/components/schemas/SkillDetailActionsResponse"
          },
          "attentionMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attentionmessage"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "displayStatus": {
            "enum": [
              "Managed",
              "Unmanaged"
            ],
            "title": "Displaystatus",
            "type": "string"
          },
          "documentMarkdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentmarkdown"
          },
          "harnessCells": {
            "items": {
              "$ref": "#/components/schemas/HarnessCellResponse"
            },
            "title": "Harnesscells",
            "type": "array"
          },
          "lineage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkillLineageInfoResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/SkillLocationResponse"
            },
            "title": "Locations",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "skillRef": {
            "title": "Skillref",
            "type": "string"
          },
          "sourceLinks": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkillSourceLinksResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "skillRef",
          "name",
          "description",
          "displayStatus",
          "attentionMessage",
          "actions",
          "harnessCells",
          "locations",
          "sourceLinks",
          "documentMarkdown"
        ],
        "title": "SkillDetailResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "skill-manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scan/skills/{skill_ref}": {
      "post": {
        "operationId": "scan_skill_api_scan_skills__skill_ref__post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ScanOptionsRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scan Skill"
      }
    },
...
    "/api/skills/id/{skill_id}/lineage": {
      "get": {
        "operationId": "get_skill_lineage_api_skills_id__skill_id__lineage_get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Skill Lineage Api Skills Id  Skill Id  Lineage Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Skill Lineage"
      }
    },
    "/api/skills/id/{skill_id}/promote": {
      "post": {
        "operationId": "promote_skill_api_skills_id__skill_id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Promote Skill Api Skills Id  Skill Id  Promote Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote Skill"
      }
    },
    "/api/skills/id/{skill_id}/restore/{version}": {
      "post": {
        "operationId": "restore_skill_version_api_skills_id__skill_id__restore__version__post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Restore Skill Version Api Skills Id  Skill Id  Restore  Version  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Skill Version"
      }
    },
    "/api/skills/id/{skill_id}/versions": {
      "get": {
        "operationId": "list_skill_versions_api_skills_id__skill_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Skill Versions Api Skills Id  Skill Id  Versions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Skill Versions"
      }
    },
...
    "/api/skills/resolve-push": {
      "post": {
        "operationId": "resolve_push_conflict_api_skills_resolve_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolvePushConflictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resolve Push Conflict Api Skills Resolve Push Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Push Conflict"
      }
    },
    "/api/skills/{skill_ref}": {
      "get": {
        "operationId": "get_skill_detail_api_skills__skill_ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Skill Detail"
      }
    },
    "/api/skills/{skill_ref}/delete": {
      "post": {
        "operationId": "delete_skill_api_skills__skill_ref__delete_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Skill"
      }
    },
    "/api/skills/{skill_ref}/disable": {
      "post": {
        "operationId": "disable_skill_api_skills__skill_ref__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableSkillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Skill"
      }
    },
    "/api/skills/{skill_ref}/enable": {
      "post": {
        "operationId": "enable_skill_api_skills__skill_ref__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableSkillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Skill"
      }
    },
    "/api/skills/{skill_ref}/manage": {
      "post": {
        "operationId": "manage_skill_api_skills__skill_ref__manage_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Manage Skill"
      }
    },
    "/api/skills/{skill_ref}/push-from-path": {
      "post": {
        "operationId": "push_skill_from_path_api_skills__skill_ref__push_from_path_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSkillFromPathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSkillResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Skill From Path"
      }
    },
    "/api/skills/{skill_ref}/resolve-conflict": {
      "post": {
        "operationId": "resolve_skill_conflict_api_skills__skill_ref__resolve_conflict_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveSkillConflictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Skill Conflict"
      }
    },
    "/api/skills/{skill_ref}/set-harnesses": {
      "post": {
        "operationId": "set_skill_harnesses_api_skills__skill_ref__set_harnesses_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSkillHarnessesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetSkillHarnessesResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Skill Harnesses"
      }
    },
    "/api/skills/{skill_ref}/source-status": {
      "get": {
        "operationId": "get_skill_source_status_api_skills__skill_ref__source_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillSourceStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Skill Source Status"
      }
    },
    "/api/skills/{skill_ref}/status-from-path": {
      "post": {
        "operationId": "skill_status_from_path_api_skills__skill_ref__status_from_path_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillStatusFromPathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillStatusResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Skill Status From Path"
      }
    },
    "/api/skills/{skill_ref}/unmanage": {
      "post": {
        "operationId": "unmanage_skill_api_skills__skill_ref__unmanage_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unmanage Skill"
      }
    },
    "/api/skills/{skill_ref}/update": {
      "post": {
        "operationId": "update_skill_api_skills__skill_ref__update_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Skill"
      }
    },
...
  }
}